requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
confusable-homoglyphs==3.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.4
django-crispy-forms==1.12.0
django-registration==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
oauthlib==3.1.1
Pillow==8.2.0
priority==1.3.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.5
zope.interface==5.4.0
